Fix specific conditions on the query
To avoid breaking existing code, the specific "where" conditions are now applied on the query rather than on the join.

This was the original behaviour before the refactor.

This bug was discovered while filtering the news (from the noviusos_news app) on a category. The inspector adds a "where" conditions in the specific conditions using the relation name as table.